Move processmanager classes to their own module and clean up the code
This add an abstract ProcessManager class to define an interface for
processmanagers, which makes it easier to use different processmanagers
interchangeably and also provides guidance for implementing new
processmanagers.
Another upside is a smaller `intelmqctl.py` file, which makes it more
readable.
The commit also refactors the code of the processmanager a bit and
makes it a more pylint conform. Exceptions are noted in the file
itself.